Fix painter retargetting issues
Changing the parent of an axis with an already existing parent was broken.
Painter retargetting using WrappedAxis was broken because WrappedAxis did not overrride clone.
Linked axes were not copying span constraints.
Removed concept of a default axis. Now axes are always passed through the GlimpseLayout's AxisFactory
(if one exists). The Axis which the user set is never used directly (although it is linked to the Axes
created for each context by default). This change has the most potential to cause unexpected problems.