Reimplement World Wind / DNC Support
The old World Wind / DNC rendering support does not work with World
Wind's WorldWindowGLJPanel. To fix this we move to using the newer
FBOGlimpseCanvas as our offscreen buffer instead of the older
SimpleOffscreenCanvas. For the most part this works. However, the
BathymetryTileExample and SurfaceTileExamples are broken. Also, the
text seems to appear blocker / more aliased. Finally, a better
heuristic to position the DNC chart drawing area in lat/lon space
is needed.