Keep miter intrustion from extending beyond the other end of the segment
There are still problems in some cases, but they no longer result in
arbitrarily large stray triangles.

AFAICT the only way to solve the remaining problems would involve
accounting for the join at endpoint A when computing the join at endpoint
B ... and vice versa. A correct solution is too complicated for a shader.
Seems like it might have to be iterative.