Speed up GLStreamingBuffer map methods with GL_MAP_INVALIDATE_RANGE_BIT
When calling glMapBufferRange, the INVALIDATE_RANGE bit tells the driver to
leave the mapped memory uninitialized. This is fine for GLStreamingBuffer,
because we need to write to the mapped memory, not read from it.

Resulting speed-up is hard to measure precisely, but certainly substantial.
For a line-strip with 10e6 vertices, glMapBufferRange with INVALIDATE_RANGE
is faster by about the time required to transfer all 10e6 vertices from
userspace to device memory.