Add support for disposing line-related objects, and improve LineProg init
LineProgram now has a constructor with no arguments, and contains an object
that holds the realized GL handles for a given GL instance. The handles are
initialized the first time begin() is called.

The handles can also be accessed directly using the new handles() method.
Calling this method will also initialize the handles, if they have not been
initialized already.