Fix a misleading comment about DNC lines and ppv-aspect-ratio
The original comment implied that handling different ppv-aspect-ratios
would be straightforward, but that was not correct -- it would be pretty
disruptive to the DNC code.

Since it is reasonable to assume that charts will be rendered with an
aspect-ratio of 1, it should be left as it is, unless we actually encounter
a compelling need for varying ppv-aspect-ratio.