Turn off aspect ratio locking in some examples
Aspect ratio locking is essentially broken when multiple aspect ratio
linked plots are linked together. The axis bounds can explode
because the x of one plot updates, making the y of the other grow a
bit. Then the y updates, growing the x even larger than it, etc...

Essentially, the axes have no sense of a desired size (i.e. the size
requested by the last api setMin/setMax call).

But just that isn't enough, the general case where x is aspect linked
to y which is aspect linked to z which is aspect linked to w is
even more complicated.

Basically all the aspect linked axes need to work together to come
to a concensus regarding the aspect ratio they should have so that
they all have min/max values as close as possible to their desired
min/max.