Add Epoch methods that use primitive-long posix-millis directly
This makes it possible to use an Epoch without creating very many TimeStamp
objects.

After lots of experience with TimeStamp, it seems cleaner in practically
all cases to use primitive-longs instead, with a convention for expressing
the units right in the variable names (e.g. "long time_PMILLIS" or "long
time_PMICROS").

In particular, with TimeStamp objects the concept of equality is especially
awkward -- but with primitive-longs it is trivially easy.