Use Vars to manage the state of LayeredGui
This is useful because other parts of the code need to react to changes in
LayeredGui's state, but we don't want everything to need a reference to the
LayeredGui. LayersPanel, for example, will now be able to operate off of a
Var<List<Layer>>.

These Vars could be pulled out into a separate model object, or into a
superclass. Not yet clear which (if either) of those would be useful.