Use "Disposable" interface with Vars, and tie up some LayeredGui loose ends
Var.addListener() now returns a Disposable instead of the listener itself
(which is a Runnable or a Consumer). The Disposable has a dispose() method
that removes the listener.

This is useful when setting up an object that adds a bunch of listeners,
and then removes them later. Rather than keeping a named reference to each
listener, we can now keep a DisposableGroup, and remove all listeners with
one line of code.