Dispose of painter asynchronously in example facets
JOGL's rules for GLAutoDrawable.invoke() are complicated and surprising,
and are less fragile when it's called asynchronously. Async is good enough
for our purposes, so it is preferrable.

There's no evidence that the synchronous call is causing the crashes, but
it wouldn't be a surprise if that turned out to be the culprit. Moreover,
the sync call was getting silently skipped some of the time, due to some
threading-related condition.