Add most of the infrastructure for handling loss of GLContext
On some platforms (especially OSX), a canvas can have its GLContext
destroyed and replaced in the course of regular usage -- e.g. when a view
is moved to a new docking location. See the explanatory comment in
GlimpseCanvasView for more info.

Still need a way to keep track of Layers that are in "limbo" -- not
currently installed, but needing to be installed the next time the view's
GLContext becomes ready.