Rework NewtSwingEDTGlimpseCanvas using AWTEDTUtil
Using AWTEDTUtil is cleaner and more effective than the previous approach.
In particular, it causes dispose() to run on the Swing thread, which is
important when canvases are being opened and closed dynamically -- as is
the case with LayeredGui.

A caveat (hopefully temporary) is that AWTEDTUtil keeps the JVM alive even
after all windows have been disposed -- it keeps doing SwingUtils.invoke()
calls, even when it has no events to dispatch. Hopefully this can be fixed
or worked around.