Add an option for GlimpseCanvasView to rely on NEWT reparenting
To use NEWT reparenting, pass "-Dlayers.glReparentingMethod=FAST" to the
JVM. This avoids destroying and recreating facets during changes to the
docking arrangement, which makes things faster and smoother -- potentially
*much* smoother, in some cases. However, on some platforms it breaks badly,
causing crashes or deadlocks.

Other recognized values are ROBUST and AUTO. The default is AUTO.

AUTO is intended to eventually choose the best reparenting method based on
the details of the current platform. However, this has not been implemented
yet, so currently AUTO always chooses ROBUST.