Reduce copying of "ViewOption..." varargs arrays
The varargs form is convenient for top-level calls, but behind the scenes
we convert the varargs array to a Collection. Previously this was done with
Arrays.asList().

However, the View superclass was then doing ImmutableSet.copyOf(), which
meant we were copying twice: once with Arrays.asList(), then again into an
ImmutableSet. But if we use ImmutableSet.copyOf() for the first copy,
instead of Arrays.asList(), then the second call to ImmutableSet.copyOf()
doesn't actually do any copying.