Catch Layer.installTo() exceptions early, log them, and carry on
In most cases it's better to pass exceptions up as high as possible before
handling them. Here, though, there is a sane policy for dealing with them
very close to the level where they get thrown, which avoids complicating a
bunch of code at higher levels.

If installTo() throws an exception, the exception gets caught and logged,
and then everything continues as if the layer had simply chosen not to
install a facet to the given view.