Iron out some wrinkles in the design of addViewPlacement()
Use new ViewPlacementRule interface for the placement chooser function.

ViewPlacementRule.getPlacement() takes two args: planArr (the arrangement
plan), and existingViewIds (a set of viewIds for which there are currently
existing views).

I initially tried providing both planArr and existingArr to the rule -- but
the rule needs to return a placement relative to nodes in planArr (because
the plan is what the rule is modifying). Passing both arrangements made it
too easy to return a placement relative to existingArr, which would later
cause an exception when applied to planArr.

Also, correlating existingArr nodes with planArr nodes is complicated, and
should probably be avoided until we actually have a use-case for it. A set
of existingViewIds seems good enough for the foreseeable future.