Add hooks in Var.set() to allow customization of listener firing
As an alternative to firing listeners inline, we can now do several Var
updates in a batch, firing the appropriate listeners only after all the
updates have finished. This allows coordinated updates of multiple Vars.

Inline and batch may be the only two dispatchers we ever need, but the door
is left open for custom ones as well.