Allow a painter to tell whether it is painting WrappedPainter's first tile
Painters often have operations they need to do once per frame (e.g. delete
GL resources that are no longer visible). This is easy in the simple case,
because once-per-frame is the same as once-per-paint.

However, WrappedPainter can call its content painters multiple times per
frame.

This change provides a way for a content painter to tell whether or not it
has already been called this frame (e.g. so that it can delete GL resources
only on the first paint of each frame).

This is a clunky way to accomplish this ... would be nice to find a cleaner
approach.