Simplify GlimpseCanvasView.glimpseInvoke()
New implemenation calls GLContext.makeCurrent() directly, instead of going
through GLAutoDrawable.invoke(). This allows us to avoid the display() call
embedded in GLAutoDrawable.invoke().

The rules about makeCurrent() and thread synchronization are complicated,
and it is possible that this new implementation violates some of those
rules. However, the situation is much more tractable if we tackle only the
simple case where all GL rendering is happening on the Swing thread.

Still need to investitage what happens if the canvas is not yet fully
realized, and possibly decide how to handle that.