Capture facet state before reparenting, and re-apply after reparenting
Currently only isVisible is captured, but this can be extended by Facet
subclasses.

The captured FacetState is not currently being serialized at any point, but
could conceivably be in the future, e.g. as a way to capture and restore
state from one run of an application to the next.