Allow MouseWrapper to work in either click-to-focus or hover-to-focus mode
It seems like click-to-focus is better if you need to click somewhere and
then type for a while. But if key presses are just used as hotkeys -- not
typing, but instantaneous actions -- then maybe hover-to-focus is better.

This commit supports focusing the hovered GlimpseLayout. To also focus the
hovered canvas, you can add a mouseEntered listener to canvas.glWindow, in
which you call canvas.glWindow.requestFocus().