Add WrappedPainter methods for managing zOrders and painter-keys
This is equivalent to the zOrder and key-based bookkeeping available in
GlimpseLayout.

It would be nice to have the same capabilities in DelegatePainter -- but
DelegatePainter does bookkeeping by "index" instead of "zOrder," which
makes it incompatible with these changes. The best option is probably to
write a new class similar to DelegatePainter, but with bookkeeping based on
zOrder and key, not on index.