Set PolygonPainter blending back the way it was
The blending issue I was seeing before was actually caused by a Background
Painter, inside a WrappedPainter, whose LookAndFeel never got set. Somehow
the WrappedPainter process of painting to a texture, and then painting the
texture, does not work well if the texture has transparent areas. Still not
sure why ... the texture definitely has an alpha channel, and I've tried
premultiplying by alpha when rendering the texture ... it's a mystery.