Set new frames visible AFTER adding all views
This is mainly intended for the case where a docker is in a modal dialog --
the EDT stops as soon as the dialog becomes visible, so we need to add all
views before that happens.

It also provides a more natural way to restore frame stacking order, by
ordering the setVisible() calls, rather than calling toFront() after the
frames are all already visible. The old way didn't work on some platforms.
This new way isn't guaranteed to either, but is more likely to work
practice.