Avoid losing views when dealing with a modal DockingGroupDialog
Views added after a modal dialog was disposed, but before the WINDOW_CLOSED
event got dispatched, would get added to the dialog being disposed.

The new behavior can still be weird -- if you add one view at a time, each
will appear, and wait for the window to be closed before moving on to the
next. But this at least makes it obvious what's happening.