Reduce flicker when moving/resizing the docking landing-indicator
When Window.setBounds() is both translating and resizing, the change is not
necessarily applied in a visually atomic way. That is, the window can
momentarily be shown with the old location and new size, or with the new
location and the old size.

This was especially noticeable with DockingGroupDialog. Views can't be
dragged out of the single window -- but the landing-indicator would
momentarily appear outside the window in some situations, which felt pretty
yucky.