Add fns for creating "submap" and "property" derived vars
A "submap" var is derived from two vars: one that holds a map, and another
that holds a set of keys. The submap var appears to hold the subset of
mappings in the map that are associated with the keys in the set.

A "property" var is derived from a single "owner" var, and appears to hold
a bean-style property of the owner object. Accessor/mutator fns specify
how to get/set the property by accessing/modifying the owner.