Check whether var value has actually changed, before running the listener
This is doesn't interact well with de-duplication in ListenableSet. This is
more important, so de-dupe has been removed.

This does an equals check at the listener level, which means equals gets
called a lot more now. But it's better than running listeners on spurious
notifications, or forcing the caller to have to think carefully about when
a caller-side check is necessary.

Fundamentally, the reason per-listener checks are required is that each
listener could potentially be hooked in through a different derived var.

An alternative approach would be to make listenable sets and derived vars
more heavy-weight, rather than simple passthroughs. That would come with
problems of its own.