Allow the caller to choose whether captureArr includes planned future views
Defaults to true -- for backward compatibility, and because that has a
significant upside when views can be added dynamically, and (usually) no
noticeable downside.

Can be set to false to prune a captured arrangement before writing it to
file. This is a bad idea if views can be added dynamically, because it will
forget the planned placements of views that haven't been added yet (but
which could be added in the future). But if the set of views is fixed, then
this will keep the saved arrangement clean by removing obsolete viewIds.