Fiddle with canvas destruction
Canvas disposal has been problematic for years -- ever since the switch to
JOGL 2, IIRC. I still don't fully understand the details ... there's a lot
of complexity in the way JOGL/NEWT makes disposal thread-safe, and a lot in
more in the interaction between NEWT and the various native impls of AWT.
The only way to solve problems in a reasonable amount of time is to try
various call sequences until we find one that works for the platforms and
situations we care about.

When using NewtSwingEDTGlimpseCanvas on Windows 10, the JVM doesn't
reliably exit after all the windows have been disposed. The proximal cause
is that the AWT-Windows thread (which is spawned by AWT's WToolkit class)
starts to receive WM_TIMER messages every 15-16 ms, and continues receiving
them indefinitely. This causes the native code in awt_Toolkit.cpp to call
AwtAutoShutdown.notifyThreadBusy(), which prevents the AWT dispatch thread
from exiting, which in turn prevents the JVM from exiting. The WM_TIMER
messages are presumably the result of a call to the Windows SetTimer()
method, but that call is not coming from the AWT native code or the
JOGL/NEWT native code, and I don't know where else to look. It may be
coming from another process, or from Windows itself for some reason.

This commit also removes the invokeLater calls from the NewtSwingEDT input
wrappers, so that event dispatch is not attempted after canvas disposal.
The invokeLater calls are unnecessary now that NewtSwingEDTGlimpseCanvas
uses AWTEDTUtil.