Split up QuickUtils methods to avoid introducing a new struct
If we provide a struct like QuickGlimpseApp, I think it will end up with
its own little API and utilities growing around it, which are usable only
with QuickGlimpseApp -- basically it will become a new API layer. I would
like to avoid doing that in QuickUtils.

I'm not necessarily against adding a similar struct in some core package,
but before we do that we should think it through thoroughly. I don't think
such a struct belongs in QuickUtils.