Put in a better workaround for the WM_TIMER problem
Unlike previous flailing, this workaround is based on an actual theory
about what is causing the issue. It's still just a theory, but it seems to
explain all the behavior to date -- including why the problem happens less
often with certain call sequences than with others -- and this commit
banishes even the most stubborn manifestation of the issue (which was
visible in VerticallyScrollableLayoutExample, by clicking the canvas and
then very quickly clicking the close-window X button).