Make the timeline EventManager a little more flexible
The getNearestRow() method, which is used when determining which event the
mouse is over, now calls the getRows() accessor method instead of using the
rows field directly. The event painter also calls the getRows() accessor --
so now the painter and the mouse handler are guaranteed to see the same row
order.

This is nice because it will allow EventManager to be extended in ways that
modify row order -- e.g. so that events stack down from the top, rather
than up from the bottom -- simply by overriding getRows().