Add support for reversing the stacking direction of timeline EventManager
This allows the "gravity" within a timeline event plot to pull events to
the top (or right) edge of the row, rather than the bottom (or left) edge.
This only matters if you have events that overlap, because that's when
stacking occurs.

To apply to all event rows in a timeline plot, override StackedTimePlot2D's
createEventPlot0() method to add the following call:

    eventPlotInfo.getEventManager( ).setReverseStackingDirection( true )