Clarify exception thrown when creating a TextureProjected2D larger than 4GB
This already threw an exception, but the exception message was confusing
because the size had overflowed, so the message said something about a
large negative number being an invalid size. Now we spot the overflow while
computing the size, and immediately throw an exception with a clear
message.

The quality of this exception message is particularly noticeable when using
Figment.