Fix resource access in several places in glimpse-core
Runtime enforcement of module boundaries in Java 9+ means that convenience
functions can no longer access arbitrary resources ... because such a
function would often be in a different jar than the resource it got asked
to access.

This means lots of code like: Foo.class.getResource("bar"). This is now
unavoidable, and should be thought of as basically a "resource literal" --
it looks like a method call, but if you try to encapsulate it you can end
up with a module-boundary violation.

Currently, the glimpse-core resource folder hierarchy doesn't quite match
the glimpse-core package hierarchy, which means we have to use absolute
resource paths. This makes resource pseudo-literals verbose and ugly. We
should consider updating the package hierarchy (while we're already making
breaking changes) so that classes can use relative resource paths.