Change GlimpseShaderProgram.addShader() arg from path-string to URL
In theory, this could have stayed with the old approach:
 1. Subclass passes in a resource path string
 2. Superclass (GlimpseShaderProgram) gets subclass using this.getClass()
 3. JOGL looks up the resource path in the context of the subclass

However, that process is subtle and confusing, and I suspect it would break
unexpectedly. Early resolution of resources into URLs (i.e. by calling
getResource on a class literal, with a string-literal argument) is less
error prone, and still pretty readable as long as relative resource paths
are used.