Handle the jaxb-runtime dependency "correctly"
This is right way to handle it ASSUMING that it gets put on the modulepath
when appropriate.

Unfortunately, Eclipse puts it on the classpath by default ... and may not
provide a good way to put it on the modulepath instead. For this reason, we
may want to file a bug report, and revert this commit until the issue has
been addressed in Eclipse.