Change version from CURRENT-SNAPSHOT to 4.0.0-SNAPSHOT, to appease javac
Maven-compiler-plugin, as of version 3.8.1, passes the Maven version to
javac with the --module-version flag. Javac then puts the string through a
set of validation rules, which were rejecting CURRENT-SNAPSHOT, causing the
compile to fail.

That seems pretty harsh, and this initially made me quite grumpy. However,
the original reason we used CURRENT-SNAPSHOT was that version ranges were
buggy (especially in Eclipse), so every time we changed the Glimpse version
we would have to then update the poms of all depent codebases -- which was
a giant pain. But version ranges seem to work fine now (including in
Eclipse), so hopefully this is okay after all.