Try moditect-maven-plugin
Gets "mvn compile" to work, and Eclipse projects are Java 8. But the reason
those work is that everything is treated as pure Java 8 ... which means
there's no build-time check for errors in the module-info files.

Also doesn't add module-info.java to source jars, which is annoying.

My final nit-pick is that the plugin execution block has to be copied to
each pom ... because if the execution block is in the parent pom, the
plugin isn't smart enough to skip the parent module, which fails.

All in all, less appealing than trying to iron out the wrinkles in one of
the earlier approaches.