Solve most issues with Java 8+ compatibility
mvn-compile and mvn-package both work. The resulting classes are all Java 8
compatible, except module-info which is Java 9. The module-info files are
checked for errors at build-time.

m2e import works. The Eclipse projects use Java 8 (which is desirable), and
ignore the module-info files (which is not ideal, but is the best Eclipse
can do at this point).

mvn-package includes the module-info source files in the jar files.

mvn-test currently does NOT seem to run any tests (but the tests do work in
Eclipse). Still investigating this one.

Modular dependents will still have a hard time depending on Glimpse Eclipse
projects in the same workspace, because the Eclipse compile won't spit out
module-info.class files. For now this seems to be an unavoidable limitation
of Eclipse. It will be annoying for people who need to work on Glimpse and
a modular dependent project at the same time. However, that probably won't
be very many people -- and even it were, this approach would still be
better than needing different Glimpse jars for different Java versions.