Expose a Runnable instead of a Cleanable in MappedFile
The standard Cleanable is in a different package depending on the Java
version, so we have to use a Glimpse-specific interface, which is non-
standard and feels clumsy.

Also, when you see "Cleanable" you start thinking about strong refs and
phantom refs. But unless you're doing something else tricky, you can use
this disposer without worrying about that stuff. (And if you ARE doing
something else tricky, the javadoc describes the behavior in enough detail
that you can worry about that stuff if you really need to.)