Remove MappedFile.readByte() and similar methods
Annoyingly, javac with "-release 8" denies access to sun.misc.Unsafe, so
the Java 8 compile step fails. This can be worked around using toolchains
... but that would require a Java 8 JDK *and* a Java 9 JDK in order to do
the build. Requiring a Java 8 JDK is reasonable, but requiring a Java 9 JDK
is over the line, because nobody actually uses Java 9. So these methods are
going away again, at least for now.