Work around another issue, to stay compatible with multiple Java versions
The removal of the built-in java.xml.bind API in Java 11 creates obstacles
to inter-version compatibility. Java 8-10 obviously can't build a Java 11
library, because of language changes. But Java 11 can't build a Java 8-10
library either, because APIs that were built into Java 8-10 are not built
into Java 11.

The glimpse-docking POM has a workaround for those obstacles, but javadoc
chokes on the workaround. So another workaround is required here, to get
javadoc to work.