Get javadoc generation working in assembly poms
Most plugin config (including the workaround for xml-bind-api being removed
from the Java 11 platform) is now inherited from the parent pom.

For now, it is difficult to get maven-javadoc-plugin to work on multiple
Jigsaw modules. Not impossible, but would be difficult and ugly. So when
generating assembly javadocs, treat sources as non-modular for the time
being.